+++ cli/cli/Commands/Profiling/RunNBomberCommand.cs
@@ -50,7 +50,7 @@ public class RunNBomberCommand : AppCommand<RunNBomberCommandArgs>
 		var prefix = args.includePrefix ? MachineHelper.GetUniqueDeviceId() : "";
 		var host = args.AppContext.Host;
 
-		var url = $"{host}/basic/{cid}.{pid}.{prefix}micro_{args.service}/{args.method}";
+		var url = $"{host}/basic/{cid}.{pid}.{args.service}/{args.method}";
 		var scope = $"{cid}.{pid}";
 
 		if (string.IsNullOrWhiteSpace(args.jsonFilePath))
@@ -69,6 +69,11 @@ public class RunNBomberCommand : AppCommand<RunNBomberCommandArgs>
 					Http.CreateRequest("POST", url)
 						.WithHeader("X-DE-SCOPE", scope)
 						.WithBody(new StringContent(jsonContent));
+				if (args.includePrefix)
+				{
+					request = request.WithHeader(Beamable.Common.Constants.Requester.HEADER_ROUTINGKEY,
+						$"{args.service}:{prefix}");
+				}
 				if (args.includeAuth)
 				{
 					request = request.WithHeader("Authorization", $"Bearer {args.AppContext.Token.Token}");