+++ cli/tests/StreamingDataTests.cs
@@ -26,10 +26,8 @@ public class StreamingDataTests
 
 		foreach (var command in commands)
 		{
-			if (command is AccountMeCommand)
-			{
-
-			}
+			if (CliGenerator.CommandTypesToReject.Contains(command.GetType())) continue;
+			
 			var commandType = command.GetType();
 			var resultStreamGenType = typeof(IResultSteam<,>);
 			var inputGenType = typeof(IHasArgs<>);